+++ src/components/home/cta-section.tsx
@@ -1,5 +1,5 @@
 import { ArrowRight, Star } from "lucide-react"
-import { Button } from "@/components/ui/button"
+import { Button } from "@/components/ui/button.tsx"
 
 export function CTASection() {
     return (
