+++ src/components/home/how-it-works-section.tsx
